refactor: completed 상위 개념으로 패키지 리팩토링
