+++ src/main/java/com/example/gimmegonghakauth/completed/controller/CompletedCoursesController.java
@@ -1,15 +1,14 @@
-package com.example.gimmegonghakauth.controller;
+package com.example.gimmegonghakauth.completed.controller;
 
-import com.example.gimmegonghakauth.domain.CompletedCoursesDomain;
+import com.example.gimmegonghakauth.completed.domain.CompletedCoursesDomain;
 import com.example.gimmegonghakauth.exception.FileException;
-import com.example.gimmegonghakauth.service.CompletedCoursesService;
+import com.example.gimmegonghakauth.completed.service.CompletedCoursesService;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
