+++ src/main/java/com/example/gimmegonghakauth/completed/infrastructure/CompletedCoursesDao.java
@@ -1,6 +1,6 @@
-package com.example.gimmegonghakauth.dao;
+package com.example.gimmegonghakauth.completed.infrastructure;
 
-import com.example.gimmegonghakauth.domain.CompletedCoursesDomain;
+import com.example.gimmegonghakauth.completed.domain.CompletedCoursesDomain;
 import com.example.gimmegonghakauth.user.domain.UserDomain;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
