+++ src/main/java/com/example/gimmegonghakauth/user/service/UserService.java
@@ -1,8 +1,8 @@
 package com.example.gimmegonghakauth.user.service;
 
-import com.example.gimmegonghakauth.dao.CompletedCoursesDao;
+import com.example.gimmegonghakauth.completed.infrastructure.CompletedCoursesDao;
 import com.example.gimmegonghakauth.user.infrastructure.UserRepository;
-import com.example.gimmegonghakauth.domain.CompletedCoursesDomain;
+import com.example.gimmegonghakauth.completed.domain.CompletedCoursesDomain;
 import com.example.gimmegonghakauth.domain.MajorsDomain;
 import com.example.gimmegonghakauth.user.domain.UserDomain;
 import com.example.gimmegonghakauth.user.service.dto.ChangePasswordDto;
