+++ src/test/java/com/example/gimmegonghakauth/dao/GonghakRepositoryTest.java
@@ -2,9 +2,10 @@ package com.example.gimmegonghakauth.dao;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import com.example.gimmegonghakauth.completed.infrastructure.CompletedCoursesDao;
 import com.example.gimmegonghakauth.constant.AbeekTypeConst;
 import com.example.gimmegonghakauth.constant.CourseCategoryConst;
-import com.example.gimmegonghakauth.domain.CompletedCoursesDomain;
+import com.example.gimmegonghakauth.completed.domain.CompletedCoursesDomain;
 import com.example.gimmegonghakauth.domain.MajorsDomain;
 import com.example.gimmegonghakauth.user.domain.UserDomain;
 import com.example.gimmegonghakauth.dto.GonghakCoursesByMajorDto;
